[#102] refactor 로그 아이템 다시 공용으로 원복, 페이지타입 프롭만 제거
